Fix carrying tensors inside loop, refactor domain concretization
