+++ include/comet/Dialect/IndexTree/Passes.h
@@ -33,8 +33,11 @@ namespace mlir
 /// Generate the code for registering conversion passes.
 #define GEN_PASS_DECL
 #include "comet/Dialect/IndexTree/Passes.h.inc"
-        /// Create a pass for the redundancy-aware kernel fusion on index tree dialect for some compound expressions
+        // Create a pass for infering the domain from the use of the index variables
         std::unique_ptr<Pass> createIndexTreeDomainInferencePass();
+
+        // Create a pass for concretizing the domain from the tensor definitions 
+        std::unique_ptr<Pass> createIndexTreeDomainConcretizationPass();
 
         /// Create a pass for applying compressed workspace transformation into IndexTreeIR
         // std::unique_ptr<Pass> createIndexTreeWorkspaceTransformationsPass();
