+++ lib/Dialect/IndexTree/CMakeLists.txt
@@ -8,6 +8,7 @@ add_mlir_dialect_library(COMETIndexTreeDialect
   # Transforms/WorkspaceTransforms.cpp 
   # Transforms/Fusion.cpp
   Transforms/IterationDomainInference.cpp
+  Transforms/DomainConcretization.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${COMET_MAIN_INCLUDE_DIR}/comet/Dialect/IndexTree
